+++ tangy-form-editor/tangy-form-condensed-editor.js
@@ -32,6 +32,9 @@ import './widget/tangy-signature-widget.js'
 import './widget/tangy-toggle-widget.js'
 import './widget/tangy-photo-capture-widget'
 import './widget/tangy-video-capture-widget'
+import './widget/tangy-audio-playback-widget.js'
+import './widget/tangy-audio-recording-widget.js'
+import './widget/tangy-audio-recording-nlp-widget.js'
 
 /**
  * `tangy-form-item-editor`