+++ tangy-form-editor/widget/tangy-audio-recording-widget.js
@@ -1,0 +1,126 @@
+import "@polymer/paper-card/paper-card.js";
+import "@polymer/paper-button/paper-button.js";
+import "tangy-form/input/tangy-audio-recording.js";
+import "tangy-form/input/tangy-checkbox.js";
+import { TangyBaseWidget } from "../tangy-base-widget.js";
+
+class TangyAudioRecordingWidget extends TangyBaseWidget {
+  get claimElement() {
+    return "tangy-audio-recording";
+  }
+
+  get defaultConfig() {
+    return {
+      ...this.defaultConfigCoreAttributes(),
+      ...this.defaultConfigQuestionAttributes(),
+      ...this.defaultConfigConditionalAttributes(),
+      ...this.defaultConfigValidationAttributes(),
+      ...this.defaultConfigAdvancedAttributes(),
+      ...this.defaultConfigUnimplementedAttributes(),
+      
+    };
+  }
+
+  upcast(config, element) {
+    return {
+      ...this.upcastCoreAttributes(config, element),
+      ...this.upcastQuestionAttributes(config, element),
+      ...this.upcastConditionalAttributes(config, element),
+      ...this.upcastValidationAttributes(config, element),
+      ...this.upcastAdvancedAttributes(config, element),
+      ...this.upcastUnimplementedAttributes(config, element),
+      ...element.getProps()
+    };
+  }
+
+  downcast(config) {
+    return `
+      <tangy-audio-recording 
+        ${this.downcastCoreAttributes(config)}
+        ${this.downcastQuestionAttributes(config)}
+        ${this.downcastConditionalAttributes(config)}
+        ${this.downcastValidationAttributes(config)}
+        ${this.downcastAdvancedAttributes(config)}
+        ${this.downcastUnimplementedAttributes(config)}
+      >
+      </tangy-audio-recording>
+    `;
+  }
+
+  renderPrint(config) {
+    return `
+    <table>
+      <tr><td><strong>Variable Name:</strong></td><td>${config.name}</td></tr>
+      <tr><td><strong>Required:</strong></td><td>${config.required}</td></tr>
+      <tr><td><strong>Disabled:</strong></td><td>${config.disabled}</td></tr>
+      <tr><td><strong>Hidden:</strong></td><td>${config.hidden}</td></tr>
+    </table>
+    <hr/>
+    `;
+  }
+
+  renderInfo(config) {
+    console.log(config)
+    const icon = (this.shadowRoot.querySelector(
+      "#icon"
+    ).innerHTML = `<span class="header-text"><mwc-icon>settings_voice</mwc-icon><span>`);
+    const name = (this.shadowRoot.querySelector(
+      "#name"
+    ).innerHTML = `<span class="header-text">${config.name}</span>`);
+    return `${icon} ${name} ${this.downcast(config)}`;
+  }
+
+  renderEdit(config) {
+    console.log(config)
+    const action = config.name ? "Edit" : "Add";
+    return `
+      <h2>${action} Audio Recording</h2>
+      
+      <tangy-form id="tangy-audio-recording">
+        <tangy-form-item>
+          <template>
+            <paper-tabs selected="0">
+                <paper-tab>General</paper-tab>
+                <paper-tab>Conditional Display</paper-tab>
+                <paper-tab>Validation</paper-tab>
+                <paper-tab>Advanced</paper-tab>
+            </paper-tabs>
+            <iron-pages selected="">
+                <div>
+                  ${this.renderEditCoreAttributes(config)}
+                  ${this.renderEditQuestionAttributes(config)}
+                </div>
+                <div>
+                  ${this.renderEditConditionalAttributes(config)}
+                </div>
+                <div>
+                  ${this.renderEditValidationAttributes(config)}
+                </div>
+                <div>
+                  ${this.renderEditAdvancedAttributes(config)}
+                </div>
+            </iron-pages>
+          </template>
+        </tangy-form-item>
+      </tangy-form>
+    `;
+  }
+
+  onSubmit(config, formEl) {
+    return {
+      ...this.onSubmitCoreAttributes(config, formEl),
+      ...this.onSubmitQuestionAttributes(config, formEl),
+      ...this.onSubmitConditionalAttributes(config, formEl),
+      ...this.onSubmitValidationAttributes(config, formEl),
+      ...this.onSubmitAdvancedAttributes(config, formEl),
+      ...this.onSubmitUnimplementedAttributes(config, formEl),
+    };
+  }
+}
+
+window.customElements.define("tangy-audio-recording-widget", TangyAudioRecordingWidget);
+window.tangyFormEditorWidgets.define(
+  "tangy-audio-recording-widget",
+  "tangy-audio-recording",
+  TangyAudioRecordingWidget
+);